🧩 Move ChangeLangMenu into SettngsModal
